Implement search currency by initials
