+++ BTC-Checker/Models/Codable.swift
