+++ BTC-Checker/Models/RestMan.swift
