+++ BTC-Checker/Views/CoinTableViewCell.swift
