+++ BTC-Checker/Views/CoinTableViewCell.xib
